pyproject.toml
[project]
name = "decontamination-demo"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.0.0",
    "elasticsearch>=9.1.0",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "spacy>=3.8.7",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.4",
]
